remove packing from drugs, add to catalog_drugs

+++ vetwebapi/vetwebapi/api_v1/drug/receipts/schemas.py
@@ -36,7 +36,7 @@ class DrugInMovementSchema(BaseModel):
     administration_method: str | None = None
     place_of_administration: str | None = None
     disposal_method: str
-    packing: int
+    packing: float
 
 
 class DrugMovementDetail(DrugMovementOut):
